extract refine_c_output
