+++ frb_codegen/src/config/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod raw_opts;
 pub(crate) mod opts;
 pub mod opts_parser;
 pub(crate) mod raw_opts_parser;
+pub(crate) mod refine_c_output;
